Develop

+++ database/migrations/2026_09_08_132141_add_type_to_chat_members.php
@@ -22,7 +22,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('chat_members', function (Blueprint $table) {
-            $table->dropColumn('type');
+            if (Schema::hasColumn('chat_members', 'type')) {
+                $table->dropColumn('type');
+            }
         });
     }
 };
